extended description explained
In this commit, I added a rough framework for checking the following to the test.py file, 3-of-a-kind, 4-of-a-kind, full house, yahtzee, and small staright. These checking cases will all certainly be changed and improved in the coming days. I removed the previous check for a large straight to find a way that works better