extended description explains commit message
In this commit, I added a start menu consisting of the following categories to help new players out, Emptyscorecard: just shows what a scorecard looks like.
Instructions: shows instructions for the game
Game rules: Shows game rules
understanding scorecard: Tells player how their points get scored.
the scorecard is a dictionary, the key being the section and the value being what the user scores. This is being put in the Yahtzee.py file because it will be a part of the game for sure with that said it will certainly be changed and bettered and made more EFFICIENT as I work through this project.